requirements.txt
bottle
pillow
chromelogger
